add __init__.py and move functionality to separate file
